Handle a VersymIndex of 0 as an error.
I noticed that the continue this patch deletes was not tested. Trying
to add a test I realized that we never put a VER_NDX_LOCAL symbol in
the dynamic symbol table. There doesn't seem to be any reason for a
linker to use VER_NDX_LOCAL for a defined shared symbol.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@320817 91177308-0d34-0410-b5e6-96231b3b80d8