[WebAssembly] When loading libraries look for companion `.imports` file
This allows libraries to supply a list of symbols which are
allowed to be undefined at link time (i.e. result in imports).

This method replaces the existing mechanism (-allow-undefined-file)
used by the clang driver to allow undefined symbols in libc.

For more on motivation for this see:
 https://github.com/WebAssembly/tool-conventions/issues/35

In the long run we hope to remove this features and instead
include this information in the object format itself.

Differential Revision: https://reviews.llvm.org/D41922

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@322320 91177308-0d34-0410-b5e6-96231b3b80d8