[tests] Fix format-binary-non-ascii.s to work with Python 3 on Windows
Some platforms interpret the pound sign as one character. Platforms that use 
Python 2.x actually interpret it as two characters because in the Python 2.x 
version of lit, the string used for the file name is a byte string and the pound 
sign is two bytes.

Patch by Stella Stamenova!


git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@329472 91177308-0d34-0410-b5e6-96231b3b80d8