Split merge sections early.
Now that getSectionPiece is fast (uses a hash) it is probably OK to
split merge sections early.

The reason I want to do this is to split eh_frame sections in the same
place.

This does mean that we have to decompress early. Given that the only
compressed sections are debug info, I don't think we are missing much.

It is a small improvement: 0.5% on the geometric mean.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@331058 91177308-0d34-0410-b5e6-96231b3b80d8