[ELF] - ICF: remove excessive check. NFC.
Change removes the excessive comparsion of
the relocation arrays sizes.

This code was dead, because at the higer level,
equalsConstant function contains the following check:

`A->NumRelocations != B->NumRelocations`
where NumRelocations contains the size of the relocations array.
So removed check did the same job twice.

This was found with use of code coverage analysis.


git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@335346 91177308-0d34-0410-b5e6-96231b3b80d8